refactor(deps): move to oclif/core as oclif/command is deprecated
