+++ package.json
@@ -6,23 +6,24 @@
   "repository": "github:linz/basemaps-config",
   "type": "module",
   "license": "MIT",
+  "main": "./build/src/index.js",
   "bin": {
-    "basemaps-config": "./bin/basemaps-config.js",
-    "basemaps-job": "./bin/basemaps-job.js"
+    "bmc": "./bin/bmc.js"
   },
   "dependencies": {
     "@basemaps/cli": "^6.14.0",
     "@basemaps/config": "^6.10.0",
     "@basemaps/geo": "^6.10.0",
     "@basemaps/shared": "^6.12.0",
-    "@chunkd/fs": "^7.3.1",
+    "@chunkd/fs": "^8.1.0",
     "@cogeotiff/chunk": "^4.2.0",
     "@linzjs/docker-command": "^6.10.0",
     "@mapbox/spritezero": "^8.0.0",
-    "@oclif/command": "^1.8.0",
+    "@oclif/core": "^1.5.3",
     "deep-diff": "^1.0.2",
-    "p-limit": "^3.1.0",
-    "pretty-json-log": "^0.3.2",
+    "p-limit": "^4.0.0",
+    "pretty-json-log": "^1.0.0",
+    "yarn": "^1.22.17",
     "zod": "^3.8.2"
   },
   "scripts": {
@@ -33,7 +34,7 @@
     "build-sprites": "node build/src/sprite.sheet.js && npx eslint build/sprites/ --fix"
   },
   "devDependencies": {
-    "@linzjs/style": "^3.1.1",
+    "@linzjs/style": "^3.7.0",
     "@types/deep-diff": "^1.0.0",
     "@types/mapbox__spritezero": "^6.3.2",
     "@types/node": "^14.11.2",
@@ -45,5 +46,9 @@
   "files": [
     "build",
     "bin"
-  ]
+  ],
+  "oclif": {
+    "commands": "./build/src/commands",
+    "bin": "bmc"
+  }
 }
